# Conflicts:
#	frontend/CretasFoodTrace/src/services/api/whitelistApiClient.ts

+++ backend-java/src/main/java/com/cretas/aims/mapper/CustomerMapper.java
@@ -59,7 +59,7 @@ public class CustomerMapper {
     /**
      * CreateRequest 转 Entity
      */
-    public Customer toEntity(CreateCustomerRequest request, String factoryId, Integer createdBy) {
+    public Customer toEntity(CreateCustomerRequest request, String factoryId, Long createdBy) {
         if (request == null) {
             return null;
         }
